+++ cmtc/src/main/scala/com/lunatech/cmt/client/Domain.scala
@@ -12,9 +12,6 @@ package com.lunatech.cmt.client
   *
   * See the License for the specific language governing permissions and limitations under the License.
   */
-
-import sbt.io.syntax.File
-
 object Domain:
   final case class ExerciseId(value: String)
   object ExerciseId:
@@ -25,14 +22,5 @@ object Domain:
   final case class TemplatePath(value: String)
   object TemplatePath:
     val default: TemplatePath = TemplatePath("")
-
-  sealed trait InstallationSource
-  object InstallationSource:
-    final case class LocalDirectory(value: File) extends InstallationSource
-    final case class ZipFile(value: File) extends InstallationSource
-    final case class GithubProject(organisation: String, project: String) extends InstallationSource {
-      val displayName = s"$organisation/$project"
-    }
-  end InstallationSource
 
 end Domain
