+++ cmtc/src/main/scala/com/lunatech/cmt/client/command/GotoExercise.scala
@@ -18,9 +18,9 @@ import com.lunatech.cmt.core.validation.Validatable
 import sbt.io.syntax.*
 import com.lunatech.cmt.client.cli.ArgParsers.{
   exerciseIdArgParser,
-  forceMoveToExerciseArgParser,
-  studentifiedRepoArgParser
+  forceMoveToExerciseArgParser
 }
+import com.lunatech.cmt.core.cli.ArgParsers.studentifiedRepoArgParser
 import com.lunatech.cmt.client.cli.CmtcCommand
 import com.lunatech.cmt.core.cli.enforceTrailingArgumentCount
