+++ cmtc/src/main/scala/com/lunatech/cmt/client/command/GotoFirstExercise.scala
@@ -6,7 +6,8 @@ import com.lunatech.cmt.{CMTcConfig, CmtError, printResult}
 import com.lunatech.cmt.client.Domain.{ExerciseId, ForceMoveToExercise}
 import com.lunatech.cmt.Domain.StudentifiedRepo
 import com.lunatech.cmt.core.validation.Validatable
-import com.lunatech.cmt.client.cli.ArgParsers.{forceMoveToExerciseArgParser, studentifiedRepoArgParser}
+import com.lunatech.cmt.client.cli.ArgParsers.forceMoveToExerciseArgParser
+import com.lunatech.cmt.core.cli.ArgParsers.studentifiedRepoArgParser
 import com.lunatech.cmt.client.cli.CmtcCommand
 import com.lunatech.cmt.core.cli.enforceNoTrailingArguments
