+++ cmtc/src/main/scala/com/lunatech/cmt/client/command/ListExercises.scala
@@ -7,7 +7,7 @@ import com.lunatech.cmt.{CMTcConfig, CmtError, printResult}
 import com.lunatech.cmt.client.cli.CmtcCommand
 import com.lunatech.cmt.core.validation.Validatable
 import com.lunatech.cmt.core.cli.enforceNoTrailingArguments
-import com.lunatech.cmt.client.cli.ArgParsers.studentifiedRepoArgParser
+import com.lunatech.cmt.core.cli.ArgParsers.studentifiedRepoArgParser
 
 object ListExercises:
