+++ cmtc/src/main/scala/com/lunatech/cmt/client/command/NextExercise.scala
@@ -16,7 +16,8 @@ import com.lunatech.cmt.client.Domain.ForceMoveToExercise
 import com.lunatech.cmt.client.command.getCurrentExerciseId
 import com.lunatech.cmt.core.validation.Validatable
 import sbt.io.syntax.*
-import com.lunatech.cmt.client.cli.ArgParsers.{forceMoveToExerciseArgParser, studentifiedRepoArgParser}
+import com.lunatech.cmt.client.cli.ArgParsers.forceMoveToExerciseArgParser
+import com.lunatech.cmt.core.cli.ArgParsers.studentifiedRepoArgParser
 import com.lunatech.cmt.client.cli.CmtcCommand
 import com.lunatech.cmt.core.cli.enforceNoTrailingArguments
