+++ cmtc/src/main/scala/com/lunatech/cmt/client/command/SetCurrentCourse.scala
@@ -2,7 +2,7 @@ package com.lunatech.cmt.client.command
 
 import caseapp.{AppName, CommandName, ExtraName, HelpMessage, RemainingArgs}
 import com.lunatech.cmt.Domain.StudentifiedRepo
-import com.lunatech.cmt.client.cli.ArgParsers.studentifiedRepoArgParser
+import com.lunatech.cmt.core.cli.ArgParsers.studentifiedRepoArgParser
 import com.lunatech.cmt.client.cli.CmtcCommand
 import com.lunatech.cmt.client.{Configuration, CurrentCourse, findStudentRepoRoot, listExercises}
 import com.lunatech.cmt.core.cli.enforceNoTrailingArguments
