+++ core/src/main/scala/com/lunatech/cmt/Domain.scala
@@ -7,4 +7,13 @@ object Domain {
   final case class StudentifiedRepo(value: File)
   object StudentifiedRepo:
     val default: StudentifiedRepo = StudentifiedRepo(file(".").getAbsoluteFile.getParentFile)
+
+  sealed trait InstallationSource
+  object InstallationSource:
+    final case class LocalDirectory(value: File) extends InstallationSource
+    final case class ZipFile(value: File) extends InstallationSource
+    final case class GithubProject(organisation: String, project: String) extends InstallationSource {
+      val displayName = s"$organisation/$project"
+    }
+  end InstallationSource
 }
